% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_afd_2022.R
\docType{data}
\name{afd_2022}
\alias{afd_2022}
\title{Party platform ("Grundsatzprogramm") of the AfD party ("Alternative for Germany") as of 2018}
\format{
A data frame containing 96 rows and 2 variables
\describe{
  \item{text}{Character. Text of the page}
  \item{page}{Integer. Page number}
}
}
\source{
AfD Deutschland, upload time stamp 2022-10-16, \url{https://www.afd.de/___grundsatzprogramm/}
}
\usage{
data(afd_2022)
}
\description{
This is a dataframe where each row is a page in the AfD party platform. 
The document is labelled to having been published in 2018, but I accessed it in 2022, so 
that's why it is called afd_2022.
}
\keyword{datasets}
