% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_afd_df.R
\docType{data}
\name{afd_df}
\alias{afd_df}
\title{Party platform ("Parteiprogramm") of the AfD party ("Alternative for Germany") as of 2016-05-01}
\format{A data frame containing 96 rows and 2 variables
\describe{
  \item{page}{int. Page number}
  \item{content}{Chr. Text of the page}
}}
\source{
\url{https://www.alternativefuer.de/wp-content/uploads/sites/7/2016/05/2016-06-27_afd-grundsatzprogramm_web-version.pdf}, by AfD Deutschland
}
\usage{
afd_df
}
\description{
Party platform ("Parteiprogramm") of the AfD party ("Alternative for Germany") as of 2016-05-01
}
\keyword{datasets}
