% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_ames_de_test.R
\docType{data}
\name{ames_de_test}
\alias{ames_de_test}
\title{Short version of the Ames housing data}
\format{
A data frame containing 500 rows and 11 columns
\describe{
  \item{id}{id}
  \item{preis}{price}
  \item{wohflaeche}{size}
  \item{grundstuecksflaeche}{lot size}
  \item{badezimmer}{number of bath rooms}
  \item{schlafzimmer}{number of rooms}
  \item{baujahr}{year built}
  \item{garage}{car garage included?}
  \item{kamin}{Fireplace included?}
  \item{umbau}{Has the house been renovated?}
  \item{pool}{Swimming pool included}
 }
}
\source{
Please cite this dataset as: 'De Cock, D. 2011. 
“Ames, Iowa: Alternative to the Boston Housing Data as an End of Semester Regression Project.” 
Journal of Statistics Education 19 (3).'
}
\usage{
data("ames_de_test")
}
\description{
This is a shortened version of the Ames housing data, test set
}
\details{
A codebook can be found here: \url{https://support.minitab.com/en-us/datasets/predictive-analytics-data-sets/ames-housing-data/} or here: \url{http://jse.amstat.org/v19n3/decock/DataDocumentation.txt}.
}
\keyword{datasets}
