% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBECT_corr_data.R
\docType{data}
\name{corr_data}
\alias{corr_data}
\title{Correlated data}
\format{
A data frame containing 900 rows and 6 columns
\describe{
  \item{V1}{Variable 1}
  \item{V2}{Variable 2}
  \item{correlationn}{correlation coefficient (Pearson)}
  \item{id}{id of the correlation (1-9)}
}
}
\usage{
data(corr_data)
}
\description{
simulated data points for two variables with varying degrees of correlation
}
\details{
The following correlations values are simulated in n=100 data points 
each: #' -0.999999, -0.90, -0.75, -0.30, 0.00, 0.30, 0.75, 0.90, 0.999999. 
Each value has a different id.
}
\keyword{datasets}
