% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_dating.R
\docType{data}
\name{dating}
\alias{dating}
\title{Results from an experiment on status symbols an dating success}
\format{A data frame containing 5063 rows and 7 columns
\describe{
  \item{strawperson}{Chr. DThere were 6 straw persons. ie., persons with a profile on the dating portal}
  \item{sex}{Chr. Sex of strawperson}
  \item{car}{Int. Whether a luxury sports car was on the profile picture or not}
  \item{date}{Chr. Date mm/dd/yyyy}
  \item{Age}{Int. Age}
  \item{ID_hash}{Chr. Username anonymized}
  \item{comm_type}{Type of communication by individuals: either matching or contact message}
  \item{whr}{Numeric. Waist hip ratio of strawperson}
  \item{beauty}{Subjective beauty rating of strawperson by rating committee}
 }}
\source{
Please cite this dataset as: 'Sauer, S., & Wolff, A. (2016, August 16). The effect of status symbols on success in online dating: an experimental study. Retrieved from osf.io/4hkjm'
}
\usage{
dating
}
\description{
Data from an online dating/mating app experiment with IV "car as status symbol"
(yes vs. no) and DV1 "matching" (number) and DV2 "message sent" (number).
The principal research question was whether a luxury sports car (BMW Z4)
on the individual profile portrait photo at an online dating app
would have an influence on the number of contact requests
and messages sent to this profile.
Care was taken that the profiles showed no difference besides the car.
}
\details{
This dataset was published here: \url{https://osf.io/4hkjm/}.
}
\keyword{datasets}
