% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_de_L.R
\docType{data}
\name{de_L}
\alias{de_L}
\title{Geo map of Germany district lines}
\format{a dataframe with 36292 rows and 4 variables
\describe{
  \item{AGZ}{num. type of border ("Art der Grenze"), 1 is state border}
\item{RDG}{num. juridical definition of border ("rechtliche Definition der Grenze"), 1 is "determined"}
\item{GM5}{num. border characteristis ("Grenzmerkmale")}
\item{geometry}{sfc_LINESTRING. geo data in shape format}
}}
\source{
\url{http://www.bkg.bund.de}
}
\usage{
de_L
}
\description{
Borders (lines) of Germany's administration districts in Shape-Format
}
\details{
licence is detailed here \url{http://www.geodatenzentrum.de/auftrag/pdf/geonutz.pdf}
}
\keyword{datasets}
