% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_parties_de.R
\docType{data}
\name{parties_de}
\alias{parties_de}
\title{Names of the parties who ran for the German federal elections 2017}
\format{
A data frame containing 43 rows and 2 columns
\describe{
  \item{party_short}{The short name (acronym) of the party}
  \item{party_long}{The long (spelled-out) name of the party}
}
}
\source{
This dataset is based on data published by the Bundeswahlleiter 2017, (c) Der Bundeswahlleiter, Wiesbaden 2017, \url{https://www.bundeswahlleiter.de/info/impressum.html}, Data is made available for unrestricted use provided the source is credited.
}
\usage{
data(parties_de)
}
\description{
Data is made available for unrestricted use provided the source is credited.
}
\keyword{datasets}
