% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_polits_twitter.R
\docType{data}
\name{polits_twitter}
\alias{polits_twitter}
\title{Dictionary of German politicians and their party affiliation}
\format{A data frame containing 200 rows and 2 variables
\describe{
  \item{party}{Chr. Short form of the affiliated party}
  \item{screenName}{Chr. Twitter account name of this politician}
}}
\source{
\url{https://twitter.com/dw_politics/lists/german-politicians?lang=en} by Deutsche Welle \url{http://www.dw.com/}
}
\usage{
data(polits_twitter)
}
\description{
This list was curated by Sebastian Sauer in August 2017. Only
politicians with active Twitter accounts were considered. Not all
politicians and not all parties where included to this list. However,
it was tried to include some of the most promiment politicians
(provided they have a Twitter account)
}
\keyword{datasets}
