% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_precip_temp_DWD.R
\docType{data}
\name{precip_temp_DWD}
\alias{precip_temp_DWD}
\title{Air temperature and precipitation in Germany}
\format{
A data frame containing 28866 rows and 5 columns
\describe{
  \item{year}{year}
  \item{month}{month}
  \item{region}{Bundesland (state)}
  \item{grundstuecksflaeche}{lot size}
  \item{precip}{Amount of precipitation in mm per square meter}
  \item{temp}{Average air temperature in degrees Celcius}
 }
}
\source{
Deutscher Wetterdienst (DWD, German Weather Service), \url{https://opendata.dwd.de/}, own elements added
}
\usage{
data("precip_temp_DWD")
}
\description{
Mean values of air temperature and precipitation across years in Germany, ad documented by the DWD
}
\details{
More information can be found at the website of eutscher Wetterdienst (DWD, German Weather Service), \url{https://dwd.de/}
}
\keyword{datasets}
