% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schimpfwoerter.R
\docType{data}
\name{schimpfwoerter}
\alias{schimpfwoerter}
\title{List of profane words (Schimpfwörter) in German}
\format{
A data frame containing 1 variable and approx. 6000 rows
\describe{
  \item{word}{Character. Profane word}
}
}
\source{
insult.wiki, \url{https://www.insult.wiki/}. Licenced under CCO.
}
\usage{
data(schimpfwoeter)
}
\description{
Schimpfwoerter collection, potentially useful for spam or hate detection. Attention: explicit content.
}
\keyword{datasets}
