% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_socec_dict.R
\docType{data}
\name{socec_dict}
\alias{socec_dict}
\title{Dataframe containing the mapping of the socio economic indicator names and their IDs (a dictionnary)}
\format{A data frame containing 52 rows and 2 columns
\describe{
  \item{ID}{chr. ID of the indicator \(short name\)}
  \item{socec_indicator}{Spelled out name of the indicator}
}}
\source{
This dataset is published by the Bundeswahlleiter 2017, (c) Der Bundeswahlleiter, Wiesbaden 2017, \url{https://www.bundeswahlleiter.de/info/impressum.html}
}
\usage{
socec_dict
}
\description{
This dataset is published by the Bundeswahlleiter 2017
(c) Der Bundeswahlleiter, Wiesbaden 2017
\url{https://www.bundeswahlleiter.de/info/impressum.html}
}
\details{
Data is made available for unrestricted use provided the source is credited.

Data presented here has been altered in the sense that it has been rendered more machine-friendly (one header row only, no special characters, no blanks in header line, etc.). Beside that, the data itself has not been changed in any way.
}
\keyword{datasets}
