% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_wildemoijs.R
\docType{data}
\name{wild_emojis}
\alias{wild_emojis}
\title{Curation of "wild" emojis}
\format{
A data frame containing 29 rows and 1 column
\describe{
  \item{emoji}{Chr. emoji}
}
}
\usage{
data(wild_emojis)
}
\description{
Several emojis frequently used to express discontent or hate
}
\details{
Subjective curation of "wild" emojis
}
\keyword{datasets}
