% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_wo_men.R
\docType{data}
\name{wo_men}
\alias{wo_men}
\title{Height, sex, and shoe size of some students.}
\format{A data frame containing 101 rows and 5 columns
\describe{
  \item{row_number}{Integer. Row number}
  \item{time}{Chr. Date and Time in German format}
  \item{sex}{(Chr. Sex; either 'woman' or 'man'}
  \item{height}{Num. Self-reported height in centimeters}
  \item{shoe_size}{Num. Self-reported shoe size in German format}
}}
\source{
The data were published here: \url{https://osf.io/ja9dw}.
   Survey conducted at the Business Psychology lab at
   FOM University of Applied Sciences from 2015 to April 2017.
   Please cite as: 'Sauer, S. (2017, February 23).
   Dataset “Height and shoe size.” http://doi.org/10.17605/OSF.IO/JA9DW'
}
\usage{
data(wo_men)
}
\description{
This dataset provides data of height, sex, and shoe size from some students.
It was collected for paedagogical purposes.
}
\keyword{datasets}
