% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_afd.R
\docType{data}
\name{afd}
\alias{afd}
\title{Party platform ("Parteiprogramm") of the AfD party ("Alternative for Germany") as of 2016-05-01}
\format{A data frame containing 96 rows and 2 variables
\describe{
  \item{page}{int. Page number}
  \item{content}{Chr. Text of the page}
}}
\source{
AfD Deutschland upload time stamp 2016-06-27
}
\usage{
data(afd)
}
\description{
This is a dataframe where each row is a page in the AfD party platform.
}
\keyword{datasets}
