% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBJECT_afd.R
\docType{data}
\name{afd}
\alias{afd}
\title{Party platform of the AfD party ("Alternative for Germany") as of 2016-05-01}
\format{A data frame containing 96 rows and variables
\describe{
  \item{page}{int. Page number}
  \item{afd_raw}{Chr. Text of the page}
}}
\source{
\url{https://www.alternativefuer.de/wp-content/uploads/sites/7/2016/05/2016-06-27_afd-grundsatzprogramm_web-version.pdf}, by AfD Deutschland
}
\usage{
afd
}
\description{
Party platform of the AfD party ("Alternative for Germany") as of 2016-05-01
}
\keyword{datasets}
